% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmeas_mm.R
\name{dmeas_mm}
\alias{dmeas_mm}
\title{Dmeas}
\usage{
dmeas_mm()
}
\arguments{
\item{N/A}{}
}
\value{
Csnippet of measurements for particle filter
}
\description{
measurement component for particle filter, written using Csnippet functionality from pomp package
}
\examples{
dmeas_mm(270.1)

}
\author{
John W. Smith Jr
}
