% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers_unde_data.R
\docType{data}
\name{drivers_unde}
\alias{drivers_unde}
\title{Meteorological driver data}
\format{
An object of class \code{"dataframe"}.
}
\source{
\href{todo}
}
\usage{
data(drivers_unde)
}
\description{
The data is from the University of Notre Dame Environmental Research Center East (UNDE). It was obtained from NEON and is used during the generation of the sythetic data
}
\examples{
data(drivers_unde)
rad <- drivers_unde$rad
}
\references{
NEON; National Ecological Observatory Network, 2020; National Ecological Observatory Network (NEON), 2022b,a
(\href{https://data.neonscience.org/}{NEON})
NEON (National Ecological Observatory Network). Shortwave radiation (primary pyranometer) (DP1.00022.001), RELEASE-2023.
(\href{https://data.neonscience.org/data-products/DP1.00022.001}{NEON Shortwave rad})
}
\keyword{datasets}
