% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpp.R
\name{GPP}
\alias{GPP}
\title{Modified GPP Function}
\usage{
GPP(lai, c_eff, maxt, mint, Ca, lat, yearday, rad)
}
\arguments{
\item{lai}{Leaf Area Index, is unitless and is the total surface of leaves per area of ground}

\item{c_eff}{Canopy Efficiency, is unitless but describes how efficiently a tree's canopy uses light}

\item{maxt}{Maximum temperature, in celsius}

\item{mint}{Minimum temperature, in celsius}

\item{Ca}{Atmospheric carbon dioxide, in parts per million}

\item{lat}{global latitude of location of observation}

\item{yearday}{which day of the year observation was recorded, as an integer}

\item{rad}{daily shortwave radiation, in gCm ^ -1}
}
\value{
GPP (gross primary productivity) from input provided
}
\description{
Creates GPP function in order to compute the GPP for a day based on values found in the drivers_unde data.
GPP function from Aggregated Canopy Model, modified for the reparameterization
of the DALEC2 model.  For use in synthetic_ym2 default particle filter.
}
\examples{
GPP(lai = 2.641, c_eff = 20, maxt = 12.83247, mint = 6.867838, Ca = 400.95, lat = 46.23, yearday = 268, rad = 11.26568)

}
\author{
John W. Smith Jr
}
