% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_ym2.R
\name{synthetic_ym2}
\alias{synthetic_ym2}
\title{Synthetic DALEC2 ecosystem simulation}
\usage{
synthetic_ym2(filter, pars, n.part = 200, scaled = TRUE, verbose = TRUE)
}
\arguments{
\item{filter}{object of class pomp, particle filter model used to generate estimate of marginal log-likelihood}

\item{pars}{named list of parameter values}

\item{n.part}{integer, number of particles (will be 200 by default)}

\item{scaled}{boolean, true if parameters are scaled to a [0,1]^10 hypercube (will be true by default)}

\item{verbose}{boolean, true if verbose warning about setting scaled to default value is needed}
}
\value{
obj: list, contains estimate of marginal log-likelihood obtained from particle filter (obj$logLik) and trajectory (obj$traj)
}
\description{
This script is designed to generate two years of synthetic dataset from a
state space formulation of the DALEC2 ecosystem model. Meteorological driver data
from the University of Notre Dame Environmental Research Center East (UNDE) were obtained
from NEON and are used during the generation of the sythetic data
}
\examples{
synthetic_ym2(synthetic_ym2_pf(), synthetic_ym2_pars(), verbose = FALSE)
synthetic_ym2(synthetic_ym2_pf(seed = 13), synthetic_ym2_pars(FALSE), n.part = 150, scaled = FALSE)

}
\author{
Turner Haugen
}
