% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_ym2_evo.R
\name{synthetic_ym2_evo}
\alias{synthetic_ym2_evo}
\title{Synthetic YM2 Evolution Function}
\usage{
synthetic_ym2_evo()
}
\arguments{
\item{N/A}{}
}
\value{
Csnippet of evolution function
}
\description{
Simulates one step forwards in time in a time series process.
}
\examples{
synthetic_ym2_evo()

}
\author{
John W. Smith Jr
}
