% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_ym2_pars.R
\name{synthetic_ym2_pars}
\alias{synthetic_ym2_pars}
\title{Synthetic YM2 default parameters}
\usage{
synthetic_ym2_pars(scaled = TRUE)
}
\arguments{
\item{scaled}{logical that decides if parameters are scaled (True) or unscaled (False)}
}
\value{
A list of parameter values
}
\description{
This script defines the parameters for synthetic_ym2, scaled to live on a [0,1]^10 hypercube.
}
\examples{
synthetic_ym2_pars()
synthetic_ym2_pars(F)

}
\author{
Turner Haugen
}
