% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_ym2_pf.R
\name{synthetic_ym2_pf}
\alias{synthetic_ym2_pf}
\title{Synthetic YM2 particle filter}
\usage{
synthetic_ym2_pf(seed = 3141)
}
\arguments{
\item{seed}{changes randomization seed, will make particle filter give different results (is 3141 by default)}
}
\value{
A Csnippet of the particle filter
}
\description{
This script defines a particle filter for use in the synthetic_ym2() function, which was designed
by John W. Smith Jr.  It embeds a DALEC2 process model for carbon evolution.
}
\examples{
filter <- synthetic_ym2_pf()
filter2 <- synthetic_ym2_pf(seed = 16766)

}
\author{
Turner Haugen
}
