% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unde_carbon_data.R
\docType{data}
\name{unde_carbon}
\alias{unde_carbon}
\title{UNDE focal forest data}
\format{
An object of class \code{"dataframe"}.
}
\source{
\href{https://doi.org/10.3334/ORNLDAAC/1567}
}
\usage{
data(unde_carbon)
}
\description{
The data is obtained from a focal forest site located at the University of Notre Dame Environmental Research Center.  It is directly used in ym2_pf().
}
\examples{
data(unde_carbon)
Lf <- unde_carbon$Lf
}
\references{
todo
}
\keyword{datasets}
