% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym2.R
\name{ym2}
\alias{ym2}
\title{DALEC2 Ecosystem With Real Data}
\usage{
ym2(filter, pars, n.part = 200)
}
\arguments{
\item{filter}{object of class pomp, contains particle filter used to model data}

\item{pars}{list of parameter values}

\item{n.part}{number of particles used in particle filter}
}
\value{
obj: list, contains estimate of marginal log-likelihood obtained from particle filter (obj$logLik) and trajectory (obj$traj)
}
\description{
todo
}
\examples{
todo

}
\author{
Turner Haugen
}
