% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym2_pars.R
\name{ym2_pars}
\alias{ym2_pars}
\title{YM2 default parameters}
\usage{
ym2_pars()
}
\arguments{
\item{N/A}{}
}
\value{
A list of parameter values
}
\description{
This script defines a default set of parameters for ym2.  Since this uses real data, these are not the TRUE parameters!  They provide the highest marginal log-likelihood of all 50000 particle MCMC iterations.
}
\examples{
ym2_pars()

}
\author{
Turner Haugen
}
