% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym2_pf.R
\name{ym2_pf}
\alias{ym2_pf}
\title{YM2 Particle Filter}
\usage{
ym2_pf()
}
\arguments{
\item{N/A}{}
}
\value{
A Csnippet of the particle filter
}
\description{
todo
}
\examples{
filter <- ym2_pf()

}
\author{
Turner Haugen
}
